requirements.txt
httpx>=0.27
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.7
pyyaml>=6.0
pandas>=2.2
openpyxl>=3.1
jinja2>=3.1
openai>=1.40
anthropic>=0.34
gitpython>=3.1
arxiv>=2.1
tqdm>=4.66
rich>=13.7
typer>=0.12
pytest>=8.2
pytest-asyncio>=0.23
